+++ flask_unchained/bundles/security/models/role.py
@@ -23,3 +23,6 @@ class Role(db.Model):
 
     def __hash__(self):
         return hash(self.name)
+
+    def __str__(self):
+        return self.name