+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - allow extensions to specify optional dependent extensions
 - add compatibility with upcoming wtforms v3
 - add support for specifying the model converter class for sqlalchemy forms
+- minor admin bundle bugfixes and improvements
 
 ## v0.9.0 (2021/06/07)